Add check for System.Object[] in git diff.

+++ app/Listeners/ProcessPullRequestReviewCommentWebhook.php
@@ -61,7 +61,9 @@ class ProcessPullRequestReviewCommentWebhook
         );
 
         // If action is deleted, we just delete the comment
-        PullRequestComment::where('id', $commentData->id)->delete();
+        if ($payload->action === 'deleted') {
+            PullRequestComment::where('id', $commentData->id)->delete();
+        }
 
         return true;
     }
